+++ src/integrationTest/integrationTestsUtilities.ts
@@ -4,28 +4,11 @@
  */
 
 import * as assert from 'assert'
+import { waitUntil } from '../shared/utilities/timeoutUtils'
 import * as vscode from 'vscode'
 
 const SECOND = 1000
 export const TIMEOUT = 30 * SECOND
-
-export async function activateExtension(extensionId: string): Promise<vscode.Extension<void>> {
-    console.log(`PID=${process.pid} activateExtension request: ${extensionId}`)
-    const extension: vscode.Extension<void> | undefined = vscode.extensions.getExtension(extensionId)
-
-    if (!extension) {
-        throw new Error(`Extension not found: ${extensionId}`)
-    }
-
-    if (!extension.isActive) {
-        console.log(`PID=${process.pid} Activating extension: ${extensionId}`)
-        await extension.activate()
-    } else {
-        console.log(`PID=${process.pid} Extension is already active: ${extensionId}`)
-    }
-
-    return extension
-}
 
 export async function sleep(miliseconds: number): Promise<void> {
     return new Promise(resolve => setTimeout(resolve, miliseconds))
@@ -34,16 +17,6 @@ export async function sleep(miliseconds: number): Promise<void> {
 // Retrieves CodeLenses from VS Code
 export async function getCodeLenses(uri: vscode.Uri): Promise<vscode.CodeLens[] | undefined> {
     return vscode.commands.executeCommand('vscode.executeCodeLensProvider', uri)
-}
-
-// Retrieves CodeLenses and asserts that undefined is not returned.
-// Convenience wrapper around the linter too.
-export async function expectCodeLenses(uri: vscode.Uri): Promise<vscode.CodeLens[]> {
-    const codeLenses = await getCodeLenses(uri)
-
-    assert.ok(codeLenses, 'Did not expect undefined when requesting CodeLenses')
-
-    return codeLenses! // appease the linter
 }
 
 export function getTestWorkspaceFolder(): string {
@@ -60,7 +33,7 @@ export function getTestWorkspaceFolder(): string {
 export async function configureAwsToolkitExtension(): Promise<void> {
     const configAws = vscode.workspace.getConfiguration('aws')
     // Prevent the extension from preemptively cancelling a 'sam local' run
-    await configAws.update('samcli.debug.attach.timeout.millis', 90000, false)
+    await configAws.update('samcli.lambda.timeout', 90000, false)
 }
 
 export async function configurePythonExtension(): Promise<void> {
@@ -100,4 +73,40 @@ export async function configureGoExtension(): Promise<void> {
     }
 
     await vscode.commands.executeCommand('go.tools.install', [gopls, dlv])
+}
+
+export async function getAddConfigCodeLens(
+    documentUri: vscode.Uri,
+    timeout: number,
+    retryInterval: number
+): Promise<vscode.CodeLens[] | undefined> {
+    return waitUntil(
+        async () => {
+            try {
+                let codeLenses = await getCodeLenses(documentUri)
+                if (!codeLenses || codeLenses.length === 0) {
+                    return undefined
+                }
+
+                // omnisharp spits out some undefined code lenses for some reason, we filter them because they are
+                // not shown to the user and do not affect how our extension is working
+                codeLenses = codeLenses.filter(codeLens => {
+                    if (codeLens.command && codeLens.command.arguments && codeLens.command.arguments.length === 3) {
+                        return codeLens.command.command === 'aws.pickAddSamDebugConfiguration'
+                    }
+
+                    return false
+                })
+
+                if (codeLenses.length > 0) {
+                    return codeLenses || []
+                }
+            } catch (e) {
+                console.log(`sam.test.ts: getAddConfigCodeLens() on "${documentUri.fsPath}" failed, retrying:\n${e}`)
+            }
+
+            return undefined
+        },
+        { timeout: timeout, interval: retryInterval, truthy: false }
+    )
 }