+++ src/integrationTest/sam.test.ts
@@ -8,6 +8,7 @@ import { Runtime } from 'aws-sdk/clients/lambda'
 import { mkdirpSync, mkdtemp, removeSync } from 'fs-extra'
 import * as path from 'path'
 import * as vscode from 'vscode'
+import * as vscodeUtils from '../../src/shared/utilities/vsCodeUtils'
 import { DependencyManager } from '../../src/lambda/models/samLambdaRuntime'
 import { helloWorldTemplate } from '../../src/lambda/models/samTemplates'
 import { getSamCliContext } from '../../src/shared/sam/cli/samCliContext'
@@ -30,7 +31,7 @@ const projectFolder = testUtils.getTestWorkspaceFolder()
 /* Test constants go here */
 const CODELENS_TIMEOUT: number = 60000
 const CODELENS_RETRY_INTERVAL: number = 200
-const DEBUG_TIMEOUT: number = 90000
+const DEBUG_TIMEOUT: number = 120000
 const NO_DEBUG_SESSION_TIMEOUT: number = 5000
 const NO_DEBUG_SESSION_INTERVAL: number = 100
 
@@ -257,38 +258,6 @@ function tryRemoveFolder(fullPath: string) {
     }
 }
 
-async function getAddConfigCodeLens(documentUri: vscode.Uri): Promise<vscode.CodeLens[] | undefined> {
-    return waitUntil(
-        async () => {
-            try {
-                let codeLenses = await testUtils.getCodeLenses(documentUri)
-                if (!codeLenses || codeLenses.length === 0) {
-                    return undefined
-                }
-
-                // omnisharp spits out some undefined code lenses for some reason, we filter them because they are
-                // not shown to the user and do not affect how our extension is working
-                codeLenses = codeLenses.filter(codeLens => {
-                    if (codeLens.command && codeLens.command.arguments && codeLens.command.arguments.length === 3) {
-                        return codeLens.command.command === 'aws.pickAddSamDebugConfiguration'
-                    }
-
-                    return false
-                })
-
-                if (codeLenses.length > 0) {
-                    return codeLenses || []
-                }
-            } catch (e) {
-                console.log(`sam.test.ts: getAddConfigCodeLens() on "${documentUri.fsPath}" failed, retrying:\n${e}`)
-            }
-
-            return undefined
-        },
-        { timeout: CODELENS_TIMEOUT, interval: CODELENS_RETRY_INTERVAL, truthy: false }
-    )
-}
-
 /**
  * Returns a string if there is a validation issue, undefined if there is no issue.
  */
@@ -387,10 +356,10 @@ async function stopDebugger(logMsg: string | undefined): Promise<void> {
 
 async function activateExtensions(): Promise<void> {
     console.log('Activating extensions...')
-    await testUtils.activateExtension(VSCODE_EXTENSION_ID.python)
-    await testUtils.activateExtension(VSCODE_EXTENSION_ID.go)
-    await testUtils.activateExtension(VSCODE_EXTENSION_ID.java)
-    await testUtils.activateExtension(VSCODE_EXTENSION_ID.javadebug)
+    await vscodeUtils.activateExtension(VSCODE_EXTENSION_ID.python, false)
+    await vscodeUtils.activateExtension(VSCODE_EXTENSION_ID.go, false)
+    await vscodeUtils.activateExtension(VSCODE_EXTENSION_ID.java, false)
+    await vscodeUtils.activateExtension(VSCODE_EXTENSION_ID.javadebug, false)
     console.log('Extensions activated')
 }
 
@@ -531,7 +500,11 @@ describe('SAM Integration Tests', async function () {
                         this.skip()
                     }
 
-                    const codeLenses = await getAddConfigCodeLens(samAppCodeUri)
+                    const codeLenses = await testUtils.getAddConfigCodeLens(
+                        samAppCodeUri,
+                        CODELENS_TIMEOUT,
+                        CODELENS_RETRY_INTERVAL
+                    )
                     assert.ok(codeLenses && codeLenses.length === 2)
 
                     let manifestFile: RegExp