+++ src/lambda/wizards/samDeployWizard.ts
@@ -28,7 +28,7 @@ import {
     WIZARD_RETRY,
 } from '../../shared/wizards/multiStepWizard'
 import { configureParameterOverrides } from '../config/configureParameterOverrides'
-import { getOverriddenParameters, getParameters } from '../utilities/parameterUtils'
+import { getOverriddenParameters, getParameters } from '../config/parameterUtils'
 import { ext } from '../../shared/extensionGlobals'
 import { EcrRepository } from '../../shared/clients/ecrClient'
 import { getSamCliVersion } from '../../shared/sam/cli/samCliContext'
@@ -43,6 +43,10 @@ import { SettingsConfiguration } from '../../shared/settingsConfiguration'
 const CREATE_NEW_BUCKET = localize('AWS.command.s3.createBucket', 'Create Bucket...')
 const ENTER_BUCKET = localize('AWS.samcli.deploy.bucket.existingLabel', 'Enter Existing Bucket Name...')
 export const CHOSEN_BUCKET_KEY = 'manuallySelectedBuckets'
+
+export interface SavedBuckets {
+    [profile: string]: { [region: string]: string }
+}
 
 export interface SamDeployWizardResponse {
     parameterOverrides: Map<string, string>
@@ -188,6 +192,47 @@ function getSingleResponse(responses: vscode.QuickPickItem[] | undefined): strin
     }
 
     return responses[0].label
+}
+
+/**
+ * The toolkit used to store saved buckets as a stringified JSON object. To ensure compatability,
+ * we need to check for this and convert them into objects.
+ */
+export function readSavedBuckets(settings: SettingsConfiguration): SavedBuckets | undefined {
+    try {
+        const buckets = settings.readSetting<SavedBuckets | string | undefined>(CHOSEN_BUCKET_KEY)
+        return typeof buckets === 'string' ? JSON.parse(buckets) : buckets
+    } catch (e) {
+        // If we fail to read settings then remove the bad data completely
+        getLogger().error('Recent bucket JSON not parseable. Rewriting recent buckets from scratch...', e)
+        settings.writeSetting(CHOSEN_BUCKET_KEY, {}, vscode.ConfigurationTarget.Global)
+        return undefined
+    }
+}
+
+/**
+ * Writes a single new saved bucket to the stored buckets setting, combining previous saved data
+ * if it exists. One saved bucket is limited per region per profile.
+ */
+export function writeSavedBucket(
+    settings: SettingsConfiguration,
+    profile: string,
+    region: string,
+    bucket: string
+): void {
+    const oldBuckets = readSavedBuckets(settings)
+
+    settings.writeSetting(
+        CHOSEN_BUCKET_KEY,
+        {
+            ...oldBuckets,
+            [profile]: {
+                ...(oldBuckets && oldBuckets[profile] ? oldBuckets[profile] : {}),
+                [region]: bucket,
+            },
+        } as SavedBuckets,
+        vscode.ConfigurationTarget.Global
+    )
 }
 
 export class DefaultSamDeployWizardContext implements SamDeployWizardContext {
@@ -621,11 +666,31 @@ export class SamDeployWizard extends MultiStepWizard<SamDeployWizardResponse> {
      */
     private hasImages: boolean = false
 
-    public constructor(private readonly context: SamDeployWizardContext, private readonly regionNode?: any) {
+    /**
+     * Initial treenode passed as a command arg to the "Deploy" command
+     * (typically, when invoked from the File Explorer context-menu).
+     */
+    private regionNode: any | undefined
+    /**
+     * Initial template.yaml path passed as a command arg to the "Deploy"
+     * command (typically, when invoked from the File Explorer context-menu).
+     */
+    private samTemplateFile: vscode.Uri | undefined
+
+    /**
+     *
+     * @param context
+     * @param commandArg Argument given by VSCode when the "Deploy" command was invoked from a context-menu.
+     */
+    public constructor(private readonly context: SamDeployWizardContext, commandArg?: any) {
         super()
-        // All nodes in the explorer should have a regionCode property, but let's make sure.
-        if (regionNode && Object.prototype.hasOwnProperty.call(regionNode, 'regionCode')) {
-            this.response.region = regionNode.regionCode
+        if (commandArg && commandArg.path) {
+            // "Deploy" command was invoked on a template.yaml file.
+            // The promptUserForSamTemplate() call will be skipped.
+            this.samTemplateFile = commandArg as vscode.Uri
+        } else if (commandArg && commandArg.regionCode) {
+            this.regionNode = commandArg
+            this.response.region = this.regionNode.regionCode
         }
     }
 
@@ -657,7 +722,17 @@ export class SamDeployWizard extends MultiStepWizard<SamDeployWizardResponse> {
     private readonly TEMPLATE: WizardStep = async () => {
         // set steps back to 0 since the next step determines if additional steps are needed
         this.context.additionalSteps = 0
-        this.response.template = await this.context.promptUserForSamTemplate(this.response.template)
+        if (this.samTemplateFile && this.response.template) {
+            // This state means:
+            //   1. wizard was started from a context-menu (`this.samTemplateFile` was set)
+            //   2. user canceled the REGION step.
+            // => User wants to exit the wizard.
+            return WIZARD_TERMINATE
+        } else if (this.samTemplateFile) {
+            this.response.template = this.samTemplateFile
+        } else {
+            this.response.template = await this.context.promptUserForSamTemplate(this.response.template)
+        }
 
         if (!this.response.template) {
             return WIZARD_TERMINATE
@@ -943,9 +1018,7 @@ async function populateS3QuickPick(
 
         let recent: string = ''
         try {
-            const bucketsJson = settings.readSetting<string | undefined>(CHOSEN_BUCKET_KEY)
-            const existingBuckets = bucketsJson ? JSON.parse(bucketsJson) : undefined
-            // JSON object of type: { [profile: string]: { [region: string]: bucket } }
+            const existingBuckets = readSavedBuckets(settings)
             if (existingBuckets && profile && existingBuckets[profile] && existingBuckets[profile][selectedRegion]) {
                 recent = existingBuckets[profile][selectedRegion]
                 baseItems.push({