+++ src/lambda/wizards/samInitWizard.ts
@@ -71,7 +71,7 @@ export interface CreateNewSamAppWizardContext {
 }
 
 export class DefaultCreateNewSamAppWizardContext extends WizardContext implements CreateNewSamAppWizardContext {
-    public readonly lambdaRuntimes = samLambdaCreatableRuntimes
+    public readonly lambdaRuntimes = samLambdaCreatableRuntimes()
     private readonly helpButton = createHelpButton(localize('AWS.command.help', 'View Toolkit Documentation'))
     private readonly currentCredentials: Credentials | undefined
     private readonly schemasRegions: Region[]