+++ src/test/shared/codelens/codeLensUtils.test.ts
@@ -124,15 +124,27 @@ describe('codeLensUtils', async function () {
                 ])
             const createSpy = sandbox.spy(Picker, 'createQuickPick')
             const finalSpy = sandbox.spy(codeLensUtils, 'pickAddSamDebugConfiguration')
-            await codeLensUtils.invokeCodeLensCommandPalette(doc, [
-                /* stub handles all pick logic */
-            ])
-            await codeLensUtils.invokeCodeLensCommandPalette(doc, [
-                /* stub handles all pick logic */
-            ])
-            await codeLensUtils.invokeCodeLensCommandPalette(doc, [
-                /* stub handles all pick logic */
-            ])
+            await codeLensUtils.invokeCodeLensCommandPalette(
+                doc,
+                [
+                    /* stub handles all pick logic */
+                ],
+                finalSpy
+            )
+            await codeLensUtils.invokeCodeLensCommandPalette(
+                doc,
+                [
+                    /* stub handles all pick logic */
+                ],
+                finalSpy
+            )
+            await codeLensUtils.invokeCodeLensCommandPalette(
+                doc,
+                [
+                    /* stub handles all pick logic */
+                ],
+                finalSpy
+            )
 
             assert.ok(createSpy.calledThrice, 'Not all test payloads run')
             assert.ok(finalSpy.notCalled, 'pickAddSamDebugConfiguration called; function did not return undefined')
@@ -167,9 +179,13 @@ describe('codeLensUtils', async function () {
                 .onFirstCall()
                 .resolves(undefined)
 
-            await codeLensUtils.invokeCodeLensCommandPalette(doc, [
-                /* stub handles all pick logic */
-            ])
+            await codeLensUtils.invokeCodeLensCommandPalette(
+                doc,
+                [
+                    /* stub handles all pick logic */
+                ],
+                finalStub
+            )
 
             assert.ok(finalStub.calledOnce, 'pickAddSamDebugConfiguration not called once and only once')
             assert.ok(finalStub.calledWith(arg0, arg1, arg2), 'pickAddSamDebugConfiguration called with incorrect args')